Refactor to be more helpful for unit testing
