+++ server.js
@@ -1,5 +1,7 @@
 var path = require('path');
 var express = require('express');
+var TimestampService = require('./timestamp-service');
+var timestampService = new TimestampService();
 var app = express();
 
 app.set('views', __dirname + '/views');
@@ -11,7 +13,7 @@ app.get('/', function(req, res) {
 
 app.get('/:timestamp', function(req, res) {
   var timestamp = req.params.timestamp;
-  var parsedResponse = getParsedResponse(timestamp);
+  var parsedResponse = timestampService.getParsedResponse(timestamp);
   var timestampObject = JSON.stringify(parsedResponse);
   res.render('timestamp', {timestampObject: timestampObject});
 })
@@ -19,19 +21,3 @@ app.get('/:timestamp', function(req, res) {
 app.listen(3000, function () {
   console.log('Timestamp Microservice listening on port 3000!');
 })
-
-function getParsedResponse(timestamp) {
-  var date = new Date(timestamp);
-  var result = {"unix": null, "natural": null};
-
-  console.log('date');
-  console.log(date);
-
-  if (date === 'Invalid Date') {
-    return result;  
-  } else {
-    result.unix = date.getTime();
-    result.natural = date.toDateString();
-    return result;
-  }
-}
